// frontend/src/api-generated/models/baoCaoTonKhoResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietTonKhoResponse } from './chiTietTonKhoResponse';

export interface BaoCaoTonKhoResponse {
  id?: number;
  idDonVi?: number;
  idViTri?: number;
  tenViTri?: string;
  idTaiSanDanhMuc?: number;
  tenTaiSanDanhMuc?: string;
  maTaiSanDanhMuc?: string;
  loaiTaiSan?: string;
  soLuongTonKho?: number;
  thoiGianCapNhat?: string;
  danhSachChiTiet?: ChiTietTonKhoResponse[];
}