// frontend/src/api-generated/models/baoCaoCapPhatResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietSuDungResponse } from './chiTietSuDungResponse';

export interface BaoCaoCapPhatResponse {
  id?: number;
  idDonVi?: number;
  idPhongBan?: number;
  tenPhongBan?: string;
  idTaiSanDanhMuc?: number;
  tenTaiSanDanhMuc?: string;
  maTaiSanDanhMuc?: string;
  loaiTaiSan?: string;
  soLuongCap?: number;
  tongGiaTriCap?: number;
  thoiGianCapNhat?: string;
  danhSachChiTiet?: ChiTietSuDungResponse[];
}