// frontend/src/api-generated/models/baoCaoBaoTriResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */
import type { ChiTietBaoTriResponse } from './chiTietBaoTriResponse';

export interface BaoCaoBaoTriResponse {
  id?: number;
  idDonVi?: number;
  idTaiSanDanhMuc?: number;
  tenTaiSanDanhMuc?: string;
  maTaiSanDanhMuc?: string;
  loaiTaiSan?: string;
  soLuong?: number;
  tongChiPhi?: number;
  tongThoiGian?: number;
  thoiGianCapNhat?: string;
  danhSachChiTiet?: ChiTietBaoTriResponse[];
}