// frontend/src/api-generated/models/baoCaoToanSanSuperAdminResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface BaoCaoToanSanSuperAdminResponse {
  idDonVi?: number;
  tenDonVi?: string;
  tongSoLuongPhanCung?: number;
  tongSoLuongPhanMem?: number;
  tongGiaTriUocTinhVnd?: number;
}